refactor out new spec-evolution-branch component
