+++ web/src/components/spec-evolution.tsx
@@ -1,20 +1,16 @@
 import React, { FunctionComponent } from 'react';
 import './spec-evolution.less';
 import {
-  Button, Header, Icon, Label, Placeholder,
+  Header, Label, Placeholder,
 } from 'semantic-ui-react';
 import {
-  ChangeProposal, EvolutionBranch, EvolutionItem, SpecItem, SpecLog, useGetInterfaceSpecEvolution,
+  SpecItem, SpecLog, useGetInterfaceSpecEvolution,
 } from '../backend-api-client';
 import { CloseSpecEvolutionButton, OpenSpecItemContentPageButton, ViewSpecLinkButton } from '../routes';
+import SpecEvolutionBranchContainer from './spec-evolution-branch';
 
 interface SpecLogItemProps {
   specItem: SpecItem;
-  interfaceName: string;
-}
-
-interface ChangeProposalProps {
-  proposedChange: ChangeProposal;
   interfaceName: string;
 }
 
@@ -32,134 +28,21 @@ const PlaceholderEvolutionBranch: FunctionComponent<SpecLogItemProps> = ({ specI
       </div>
     </div>
     <div key={specItem.ref} className="item">
-      <LatestAgreedLogItem specItem={specItem} interfaceName={interfaceName} />
-    </div>
-  </div>
-);
-
-const ChangeProposalItem: FunctionComponent<ChangeProposalProps> = ({ proposedChange, interfaceName }) => (
-  <div className="log-entry-container" data-testid="log-entry-container">
-    <div className="line-container">
-      <div className="latest-agreed line" />
-      <div className="change-proposal line" />
-    </div>
-    <div className="details-container">
-      <Button
-        icon
-        labelPosition="right"
-        size="mini"
-        color="green"
-        href={proposedChange.pullRequest.url}
-        target="_blank"
-        rel="noopener noreferrer"
-      >
-        PR #
-        {proposedChange.pullRequest.number}
-        <Icon name="code branch" />
-      </Button>
-      <div className="centre">{proposedChange.pullRequest.title}</div>
-      <OpenSpecItemContentPageButton specItem={proposedChange.specItem} />
-      <ViewSpecLinkButton refName={proposedChange.specItem.ref} interfaceName={interfaceName} />
-    </div>
-  </div>
-);
-
-const LatestAgreedLogItem: FunctionComponent<SpecLogItemProps> = ({ specItem, interfaceName }) => (
-  <div className="log-entry-container" data-testid="log-entry-container">
-    <div className="line-container">
-      <div className="latest-agreed line" />
-    </div>
-    <div className="details-container">
-      <Label color="blue">{specItem.ref}</Label>
-      <Label color="blue" tag>{specItem.parseResult.openApiSpec.version}</Label>
-      <div className="centre" />
-      <OpenSpecItemContentPageButton specItem={specItem} />
-      <ViewSpecLinkButton refName={specItem.ref} interfaceName={interfaceName} />
-    </div>
-  </div>
-);
-
-interface EvolutionItemProps {
-  evolutionItem: EvolutionItem;
-  isMain: boolean;
-}
-
-const EvolutionItemLines: FunctionComponent<EvolutionItemProps> = ({ evolutionItem, isMain }) => {
-  const { pullRequest } = evolutionItem;
-
-  const latestAgreedLine = isMain ? (<div className="latest-agreed line" />) : (<div className="line" />);
-  const upcomingReleaseLine = !isMain ? (<div className="upcoming-release line" />) : null;
-  const pullRequestLine = pullRequest ? (<div className="change-proposal line" />) : null;
-
-  return (
-    <div className="line-container">
-      {latestAgreedLine}
-      {upcomingReleaseLine}
-      {pullRequestLine}
-    </div>
-  );
-};
-
-const EvolutionItemDetails: FunctionComponent<EvolutionItemProps> = ({ evolutionItem, isMain }) => {
-  const { pullRequest, tag } = evolutionItem;
-
-  const tagLabel = tag
-    ? (<Label color="blue" tag>{tag}</Label>)
-    : null;
-
-  let prLabel = null;
-  let prTitle = null;
-  if (evolutionItem.pullRequest) {
-    prLabel = (
-      <Button
-        icon
-        labelPosition="right"
-        size="mini"
-        color="green"
-        href={pullRequest.url}
-        target="_blank"
-        rel="noopener noreferrer"
-      >
-        PR #
-        {pullRequest.number}
-        <Icon name="code branch" />
-      </Button>
-    );
-    prTitle = (<div className="centre">{pullRequest.title}</div>);
-  }
-
-  return (
-    <div className="details-container">
-      {tagLabel}
-      {prLabel}
-      {prTitle}
-    </div>
-  );
-};
-
-const buildEvolutionBranch = (evolutionBranch: EvolutionBranch, isMain: boolean) => {
-  const { evolutionItems } = evolutionBranch;
-  const logItems = evolutionItems.map((evolutionItem) => (
-    <div className="item">
       <div className="log-entry-container" data-testid="log-entry-container">
-        <EvolutionItemLines evolutionItem={evolutionItem} isMain={isMain} />
-        <EvolutionItemDetails evolutionItem={evolutionItem} isMain={isMain} />
-        {/* <div className="details-container">
-          <Label color="blue">{evolutionItem}</Label>
+        <div className="line-container">
+          <div className="latest-agreed line" />
+        </div>
+        <div className="details-container">
+          <Label color="blue">{specItem.ref}</Label>
           <Label color="blue" tag>{specItem.parseResult.openApiSpec.version}</Label>
           <div className="centre" />
           <OpenSpecItemContentPageButton specItem={specItem} />
           <ViewSpecLinkButton refName={specItem.ref} interfaceName={interfaceName} />
-        </div> */}
+        </div>
       </div>
     </div>
-  ));
-  return (
-    <div className="evolution-branch-container">
-      {logItems}
-    </div>
-  );
-};
+  </div>
+);
 
 interface SpecLogProps {
   specLog: SpecLog;
@@ -170,15 +53,15 @@ interface SpecLogProps {
 const SpecEvolutionContainer: FunctionComponent<SpecLogProps> = ({ specLog, interfaceName, encodedId }) => {
   const getInterfaceSpecEvolution = useGetInterfaceSpecEvolution({ encodedId, interfaceName });
 
-  const { data: interfaceSpecEvolutionResult, loading, error } = getInterfaceSpecEvolution;
+  const { data: interfaceSpecEvolutionResult, loading } = getInterfaceSpecEvolution;
 
   let evolutionBranches = null;
   if (loading) {
     evolutionBranches = [(<PlaceholderEvolutionBranch specItem={specLog.latestAgreed} interfaceName={interfaceName} />)];
   } else {
     const { main, releases } = interfaceSpecEvolutionResult.specEvolution;
-    const mainEvolutionBranch = buildEvolutionBranch(main, true);
-    const releaseEvolutionBranches = releases.map((releaseBranch) => buildEvolutionBranch(releaseBranch, false));
+    const mainEvolutionBranch = (<SpecEvolutionBranchContainer evolutionBranch={main} isMain />);
+    const releaseEvolutionBranches = releases.map((releaseBranch) => (<SpecEvolutionBranchContainer evolutionBranch={releaseBranch} />));
     evolutionBranches = [...releaseEvolutionBranches, mainEvolutionBranch];
   }
